refactor directories and filenames
